Task2

+++ Excercises/d2ch/Task1.java
@@ -4,17 +4,17 @@ import java.util.Random;
 import java.awt.geom.Point2D;
 
 public class Task1 {
-    public static float GetOrientation_1(Point2D.Float p1, Point2D.Float p2, Point2D.Float p3) {
+    public static double GetOrientation_1(Point2D.Double p1, Point2D.Double p2, Point2D.Double p3) {
 
         return p1.x*(p2.y - p3.y) + p2.x*(p3.y - p1.y) + p3.x*(p1.y - p2.y);
     }
 
-    public static float GetOrientation_2(Point2D.Float p1, Point2D.Float p2, Point2D.Float p3) {
+    public static double GetOrientation_2(Point2D.Double p1, Point2D.Double p2, Point2D.Double p3) {
 
         return p1.x*p2.y - p1.x*p3.y + p2.x*p3.y - p2.x*p1.y + p3.x*p1.y - p3.x*p2.y;
     }
 
-    public static float GetOrientation_3(Point2D.Float p1, Point2D.Float p2, Point2D.Float p3) {
+    public static double GetOrientation_3(Point2D.Double p1, Point2D.Double p2, Point2D.Double p3) {
 
         return p1.x*p2.y + p2.x*p3.y + p3.x*p1.y - p1.x*p3.y - p2.x*p1.y - p3.x*p2.y;
     }
@@ -23,12 +23,12 @@ public class Task1 {
         int n = (int)Math.pow(10, 5);
         
         var random = new Random(6);
-        Point2D.Float[] array = new Point2D.Float[n];
-        float x = 0;
+        Point2D.Double[] array = new Point2D.Double[n];
+        double x = 0;
         for(int i = 0; i < n; i++) 
         {
-            array[i] = new Point2D.Float(x, x*x);
-            x += random.nextFloat();
+            array[i] = new Point2D.Double(x, x*x);
+            x += random.nextDouble();
         }
 
         int errors = 0;
